backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
python-dotenv>=1.0
psycopg[binary]>=3.1
dj-database-url>=2.2
whitenoise>=6.7
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
Pillow>=10.0
gunicorn>=21.2
